added tests, added way to override name
