+++ Cargo.toml
@@ -24,6 +24,7 @@ members = [
   "state-chain/chains",
   "state-chain/custom-rpc",
   "state-chain/generic-typeinfo-derive",
+  "state-chain/generic-typeinfo-derive/macro",
   "state-chain/node",
   "state-chain/pallets/cf-account-roles",
   "state-chain/pallets/cf-asset-balances",
@@ -293,6 +294,7 @@ engine-upgrade-utils = { path = "engine-upgrade-utils" }
 multisig = { path = "engine/multisig" }
 scale-json = { path = "utilities/scale-json", default-features = false }
 sol-prim = { path = "foreign-chains/solana/sol-prim", default-features = false }
+generic-typeinfo-derive = { path = "state-chain/generic-typeinfo-derive/macro" }
 
 # The list of dependencies below (which can be both direct and indirect dependencies) are crates
 # that are suspected to be CPU-intensive, and that are unlikely to require debugging (as some of
