+++ state-chain/generic-typeinfo-derive/macro/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "generic-typeinfo-derive"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+proc-macro2 = "1.0.89"
+quote = { workspace = true }
+syn = { workspace = true }
+
+[lints]
+workspace = true
+
+[lib]
+proc-macro = true
