+++ state-chain/pallets/cf-broadcast/Cargo.toml
@@ -22,7 +22,7 @@ cf-primitives = { workspace = true }
 cf-traits = { workspace = true }
 cfe-events = { workspace = true }
 cf-runtime-utilities = { workspace = true }
-generic-typeinfo-derive = { path = '../../generic-typeinfo-derive'}
+generic-typeinfo-derive = { workspace = true }
 
 log = { workspace = true }
 serde = { workspace = true, features = ["derive", "alloc"] }
