+++ state-chain/pallets/cf-elections/Cargo.toml
@@ -25,7 +25,7 @@ cf-primitives = { workspace = true }
 # Utility deps
 cf-utilities = { workspace = true }
 cf-runtime-utilities = { workspace = true }
-generic-typeinfo-derive = { path = '../../generic-typeinfo-derive' }
+generic-typeinfo-derive = { workspace = true }
 
 # Non-substrate external deps
 bitvec = { workspace = true, features = ["alloc"] }
