+++ state-chain/pallets/cf-ingress-egress/Cargo.toml
@@ -18,7 +18,7 @@ cf-chains = { workspace = true }
 cf-primitives = { workspace = true }
 cf-traits = { workspace = true }
 cf-runtime-utilities = { workspace = true, features = ["derive"] }
-generic-typeinfo-derive = { path = '../../generic-typeinfo-derive' }
+generic-typeinfo-derive = { workspace = true }
 
 log = { workspace = true }
 strum = { workspace = true }
